chore: move security configs to its own package
